requirements.txt
selenium
requests
bs4
discord.py